Add new permission rules

+++ src/modules/tasks/tasks.service.ts
@@ -1,7 +1,7 @@
 import { Tasks as Task } from './tasks.model';
-import { Groups as Group} from '../groups/groups.model';
+import { Groups as Group } from '../groups/groups.model';
 import { CustomUser } from '../../common/types/types';
-import { NotFound, Forbidden} from '../../common/exeptions/';
+import { NotFound, Forbidden } from '../../common/exeptions/';
 import { sequelize } from '../../database';
 
 interface ITasks {
@@ -13,55 +13,47 @@ interface ITasks {
 const NO_RIGHTS = 'You do not have rights to do this.';
 
 class TasksService {
-  public async findByGroup(user: CustomUser): Promise<Task[]> {
-    if (!user.isMentor) {
-      const tasks: Task[] = await Task.findAll({ where: { groupId: user.groupId } });
+  public async findAll(user: CustomUser): Promise<Task[]> {
+    if (!user.isMentor) { 
+      const tasks: Task[] = await Task.findAll({ where: {groupId: user.groupId} });
       if (!tasks.length) {
-        throw new NotFound(`There is no tasks for group with groupId ${user.groupId}`)
-      }
+        throw new NotFound(`There is no tasks for group with groupId ${user.groupId}`);
+      } 
       return tasks;
     }
+
     return await Task.findAll();
   }
 
   public async findOneById(id: number, user: CustomUser) {
-    const task: Task = await Task.findOne({ where: { id }});
+    const task: Task = await Task.findOne({ where: { id } });
     if (!task) {
       throw new NotFound(`Can't find task with id ${id}`);
     }
 
-    if (!user.isMentor) {
-      if (task.groupId === user.groupId) {
-        return task
-      } else {
-        throw new Forbidden(NO_RIGHTS);
-      }
-    } 
+    if (!user.isMentor && user.groupId !== task.groupId) {
+      throw new Forbidden(NO_RIGHTS);
+    }
+
     return task;
   }
 
-  public async createOne(
-    task: ITasks,
-    user: CustomUser
-  ): Promise<Task> {
+  public async createOne(task: ITasks, user: CustomUser): Promise<Task> {
     if (!user.isMentor) {
       throw new Forbidden(NO_RIGHTS);
     }
 
-    if (user.isMentor && !user.isAdmin) {
-      return sequelize.transaction(async (transaction) => {
-        const group: Group = await Group.findOne({where: {id: task.groupId}, transaction})
-        if (group.teacherId !== user.id) {
-          throw new Forbidden(NO_RIGHTS);
-        } else {
-          return await Task.create(task, { transaction });
-        }
+    return sequelize.transaction(async (transaction) => {
+      const group: Group = await Group.findOne({
+        where: { id: task.groupId },
+        transaction,
       });
-    } 
+      if (user.isMentor && !user.isAdmin && user.id !== group.teacherId) {
+        throw new Forbidden(NO_RIGHTS);
+      }
 
-    if (user.isAdmin) {
-      return await Task.create(task);
-    }
+      return await Task.create(task, { transaction });
+    });
   }
 
   public async updateOne(
@@ -79,10 +71,18 @@ class TasksService {
         throw new NotFound(`Can't find task with id ${id}`);
       }
 
+      const group: Group = await Group.findOne({
+        where: { id: task.groupId },
+        transaction,
+      });
+      if (user.isMentor && !user.isAdmin && user.id !== group.teacherId) {
+        throw new Forbidden(NO_RIGHTS);
+      }
+
       const [updatedRow, [updatedTask]] = await Task.update(updates, {
         returning: true,
         where: { id },
-        transaction
+        transaction,
       });
       return updatedTask;
     });
@@ -97,6 +97,14 @@ class TasksService {
       const task: Task = await Task.findOne({ where: { id }, transaction });
       if (!task) {
         throw new NotFound(`Can't find task with id ${id}`);
+      }
+
+      const group: Group = await Group.findOne({
+        where: { id: task.groupId },
+        transaction,
+      });
+      if (user.isMentor && !user.isAdmin && user.id !== group.teacherId) {
+        throw new Forbidden(NO_RIGHTS);
       }
 
       await Task.destroy({ where: { id }, transaction });
